requirements.txt
fastapi==0.101.1
fastapi-mail==1.4.1
uvicorn==0.23.2
gunicorn==21.2.0
pytest==7.4.0
pytest-mock==3.11.1
psycopg-binary==3.1.10
psycopg2==2.9.7
python-jose==3.3.0
python-multipart==0.0.6
starlette~=0.27.0
requests~=2.31.0
pydantic==2.3.0
pyotp~=2.9.0
pycryptodome==3.18.0
pyhibp==4.2.0